feat: add existing sqlx.DB support
allows passing of an existing sqlx.DB object to create persisters. this
allows us to use a single pool and instance to produce all the required
persister types.  sort of weird, probably should get rid of the
separated type model and just use a single persister type, but let's just leave it for now